+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Mobile, Web: Renamed 'Projects' page to 'Tags', added contexts to it.
 - Mobile, Web: Not showing empty task when creating new todo.txt file.
 - Mobile, Web: Sorting tasks by project first, then by context.
+- Mobile, Web: Added support for 'color' extension.
 
 ## 1.1.0